Initial push
Pushed all of the files for both the graphic UI (p5.js) and the start of the move analyzer, skunk. There are also two other weak javascript bots one with minmax and alpha beta pruning and the other with a Monte Carlo search (both are very weak and only search 3-4 ply).