Fixed 3 fold repetition + UCI promotion parse move error
Found out that there is an option to send move history to the engine from the UCI. This is important for avoiding 3 fold repetition. I also increased the negative penalty that this has. Also found an error where the GUI sends a history move to the engine and it parsed it incorrectly, causing an error in the GUI. This has been fixed. Enabled futility pruning, it seems to play ok.